ミュート関係の設定項目を/settings/mute-blockにまとめる
