+++ locales/en-US.yml
@@ -1114,7 +1114,7 @@ _wordMute:
   muteWordsDescription: "Separate with spaces for an AND condition or with line breaks for an OR condition."
   muteWordsDescription2: "Surround keywords with slashes to use regular expressions."
   softDescription: "Hide notes that fulfil the set conditions from the timeline."
-  hardDescription: "Prevents notes fulfilling the set conditions from being added to the timeline. In addition, these notes will not be added to the timeline even if the conditions are changed."
+  hardDescription: "This feature is auxiliary and will not work under some conditions. Prevents notes fulfilling the set conditions from being added to the timeline. In addition, these notes will not be added to the timeline even if the conditions are changed."
   soft: "Soft"
   hard: "Hard"
   mutedNotes: "Muted notes"
