+++ locales/ja-JP.yml
@@ -1119,8 +1119,8 @@ _wordMute:
   muteWords: "ミュートするワード"
   muteWordsDescription: "スペースで区切るとAND指定になり、改行で区切るとOR指定になります。"
   muteWordsDescription2: "キーワードをスラッシュで囲むと正規表現になります。"
-  softDescription: "指定した条件のノートをタイムラインから隠します。"
-  hardDescription: "指定した条件のノートをタイムラインに追加しないようにします。追加されなかったノートは、条件を変更しても除外されたままになります。"
+  softDescription: "指定した単語を含むノートをタイムラインから隠します。処理はクライアントサイドで行われるため、サードパーティクライアント使用時には適用されません。"
+  hardDescription: "この機能は補助的なものであり、条件によっては動作しません。指定した単語を含むノートをタイムラインに追加しないようにします。追加されなかったノートは、条件を変更しても除外されたままになります。処理はバックエンドで行われるため、サードパーティクライアントにも適用されます。"
   soft: "ソフト"
   hard: "ハード"
   mutedNotes: "ミュートされたノート"
