+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "misskey",
-	"version": "12.121.6-b2",
+	"version": "12.121.6-b3",
 	"codename": "agathos",
 	"repository": {
 		"type": "git",
